Change plan to use script
