+++ .circleci/config.yml
@@ -40,6 +40,8 @@ jobs:
         run: "npm install"
       -
         run: "cp ~/repo/api/.keys.json ~/repo/api/keys.json"
+      -
+        run: "npm run db_init"
       - 
         run: "npm run test"
       -
