+++ package.json
@@ -5,6 +5,7 @@
   "main": "api/index.js",
   "scripts": {
     "start": "node api/index.js",
+    "db_init": "node ./db/db_init.js",
     "test": "NODE_ENV='test' ./node_modules/mocha/bin/mocha"
   },
   "repository": {
