--- test/mocha.opts
@@ -1,2 +1,0 @@
---recursive
-test/test_helpers/db_circle_init.js
